requirements.txt
python-dotenv
pandas
numpy
pymongo
Certifi
pymongo
pymongo[srv]==3.6
# -e .

        